requirements.txt
orionsdk
python-dotenv
requests-mock